Remove label-break from updates, it's too hard.
While in principle, one can define the behaviour of label ... break
on the left-hand side of update expressions,
it requires an adaptation of the operators v[] |= f et al.,
things like `first(.[]) += 1` do not compose correctly.
This becomes messy quite fast and also has the disadvantage
that evaluation order for objects becomes undefined.
You can already see that from the behaviour of
`{a: 0, b: 0} | first(.[]) += 1` --- does this update a or b?
Furthermore, just forbidding label-break in path expressions
simplifies the definition of the break() function
(it does not have to take a value anymore).
Also, that way, we do not have to wrestle with polarities
already in the "Values" section.

TL;DR: label-break in path expressions is a bad idea.